chore(backend): Move webhook types to their own file
